requirements.txt
pyyaml
click
click-config-file